pyproject.toml
[project]
name = "rag-bot"
version = "0.0.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.4.1",
    "fastapi>=0.128.0",
    "llama-index>=0.14.13",
    "llama-index-embeddings-openai>=0.5.1",
    "llama-index-llms-openai>=0.6.13",
    "llama-index-retrievers-bm25>=0.6.5",
    "llama-index-vector-stores-chroma>=0.5.5",
    "openai>=2.15.0",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "rank-bm25>=0.2.2",
    "tqdm>=4.67.1",
    "uvicorn>=0.40.0",
]
